requirements.txt
disnake==2.10.1
python-dotenv==1.1.0
